kotlin sucks

+++ src/main/java/net/ccbluex/liquidbounce/launch/data/legacyui/GuiMainMenu.kt
@@ -15,6 +15,7 @@ import net.minecraft.client.gui.*
 import net.minecraft.client.resources.I18n
 import net.minecraftforge.fml.client.GuiModList
 import java.awt.Color
+import org.lwjgl.input.Mouse
 
 class GuiMainMenu : GuiScreen(), GuiYesNoCallback {
     override fun initGui() {
@@ -50,7 +51,7 @@ class GuiMainMenu : GuiScreen(), GuiYesNoCallback {
             RenderUtils.drawRect(start, 15f, start + mc.fontRendererObj.getStringWidth(str), 15f + mc.fontRendererObj.FONT_HEIGHT, Color.BLACK.rgb)
             mc.fontRendererObj.drawString(str, start, 15f, Color.WHITE.rgb, false)
         }
-        Gui.drawModalRectWithCustomSizedTexture(-21 + Mouse.getX() / 90, Mouse.getY() * -1 / 90, 0.0f, 0.0f, this.width + 20, this.height + 20, (float)(this.width + 21), (float)(this.height + 20));
+        Gui.drawModalRectWithCustomSizedTexture(-21 + Mouse.getX() / 90, Mouse.getY() * -1 / 90, 0.0f, 0.0f, this.width + 20, this.height + 20, (this.width + 21), (this.height + 20));
         super.drawScreen(mouseX, mouseY, partialTicks)
     }
